Fix memory leak + better separation of responsibilities
